+++ src/modules/esm.mjs
@@ -1,0 +1,52 @@
+import { readFile } from 'fs/promises';
+import path from 'path';
+import { fileURLToPath } from 'url';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import './files/c.js';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const random = Math.random();
+
+const aJson = JSON.parse(
+    await readFile(
+        new URL('./files/a.json', import.meta.url)
+    )
+);
+
+const bJson = JSON.parse(
+    await readFile(
+        new URL('./files/b.json', import.meta.url)
+    )
+);
+
+let randomObject;
+
+if (random > 0.5) {
+    randomObject = aJson;
+} else {
+    randomObject = bJson;
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(randomObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { randomObject, myServer };
